+++ src/main/java/uk/gov/hmcts/reform/prl/handlers/CaseEventHandler.java
@@ -11,6 +11,7 @@ import uk.gov.hmcts.reform.prl.enums.YesOrNo;
 import uk.gov.hmcts.reform.prl.enums.c100respondentsolicitor.RespondentSolicitorEvents;
 import uk.gov.hmcts.reform.prl.enums.noticeofchange.SolicitorRole;
 import uk.gov.hmcts.reform.prl.events.CaseDataChanged;
+import uk.gov.hmcts.reform.prl.handlers.service.CaseEventHandlerService;
 import uk.gov.hmcts.reform.prl.models.Element;
 import uk.gov.hmcts.reform.prl.models.EventValidationErrors;
 import uk.gov.hmcts.reform.prl.models.c100respondentsolicitor.RespondentEventValidationErrors;
@@ -56,22 +57,18 @@ public class CaseEventHandler {
     public static final String ID = "id";
 
     private final CoreCaseDataService coreCaseDataService;
-    private final TaskListService taskListService;
-    private final TaskListRenderer taskListRenderer;
-    private final RespondentSolicitorTaskListRenderer respondentSolicitorTaskListRenderer;
-    private final TaskErrorService taskErrorService;
-    private final RespondentTaskErrorService respondentTaskErrorService;
+    private final CaseEventHandlerService caseEventHandlerService;
 
     @EventListener
     public void handleCaseDataChange(final CaseDataChanged event) {
         final CaseData caseData = event.getCaseData();
 
-        final String taskList = getUpdatedTaskList(caseData);
-        final String respondentTaskListA = getRespondentTaskList(caseData, C100_RESPONDENT_EVENTS_A);
-        final String respondentTaskListB = getRespondentTaskList(caseData, C100_RESPONDENT_EVENTS_B);
-        final String respondentTaskListC = getRespondentTaskList(caseData, C100_RESPONDENT_EVENTS_C);
-        final String respondentTaskListD = getRespondentTaskList(caseData, C100_RESPONDENT_EVENTS_D);
-        final String respondentTaskListE = getRespondentTaskList(caseData, C100_RESPONDENT_EVENTS_E);
+        final String taskList = caseEventHandlerService.getUpdatedTaskList(caseData);
+        final String respondentTaskListA = caseEventHandlerService.getRespondentTaskList(caseData, C100_RESPONDENT_EVENTS_A);
+        final String respondentTaskListB = caseEventHandlerService.getRespondentTaskList(caseData, C100_RESPONDENT_EVENTS_B);
+        final String respondentTaskListC = caseEventHandlerService.getRespondentTaskList(caseData, C100_RESPONDENT_EVENTS_C);
+        final String respondentTaskListD = caseEventHandlerService.getRespondentTaskList(caseData, C100_RESPONDENT_EVENTS_D);
+        final String respondentTaskListE = caseEventHandlerService.getRespondentTaskList(caseData, C100_RESPONDENT_EVENTS_E);
 
         coreCaseDataService.triggerEvent(
             JURISDICTION,
@@ -97,80 +94,6 @@ public class CaseEventHandler {
                 String.valueOf(caseData.getId())
             )
         );
-    }
-
-    public String getUpdatedTaskList(CaseData caseData) {
-        String taskList = "";
-        if (caseData.getState() != null
-            && (caseData.getState().equals(State.AWAITING_SUBMISSION_TO_HMCTS)
-            || caseData.getState().equals(State.AWAITING_RESUBMISSION_TO_HMCTS))) {
-            final List<Task> tasks = taskListService.getTasksForOpenCase(caseData);
-            List<EventValidationErrors> eventErrors = taskErrorService.getEventErrors(caseData);
-            if (caseData.getCaseTypeOfApplication().equalsIgnoreCase(C100_CASE_TYPE)) {
-                List<Event> events = taskListService.getC100Events(caseData);
-                eventErrors.removeIf(e -> !events.contains(e.getEvent()));
-            }
-
-            if (caseData.getCaseTypeOfApplication().equalsIgnoreCase(FL401_CASE_TYPE)) {
-                List<Event> events = taskListService.getFL401Events(caseData);
-                eventErrors.removeIf(e -> !events.contains(e.getEvent()));
-            }
-
-            taskList = taskListRenderer
-                .render(
-                    tasks,
-                    eventErrors,
-                    caseData.getCaseTypeOfApplication().equalsIgnoreCase(C100_CASE_TYPE),
-                    caseData
-                );
-        }
-        return taskList;
-    }
-
-    public String getRespondentTaskList(CaseData caseData, String respondent) {
-        String respondentTaskList = "";
-        if (caseData.getRespondents() != null
-            && !caseData.getRespondents().isEmpty()) {
-            Optional<SolicitorRole> solicitorRole = SolicitorRole.from(respondent);
-            if (solicitorRole.isPresent() && caseData.getRespondents().size() > solicitorRole.get().getIndex()) {
-                Element<PartyDetails> respondingParty = caseData.getRespondents().get(solicitorRole.get().getIndex());
-                if (respondingParty.getValue() != null
-                    && respondingParty.getValue().getUser() != null
-                    && YesOrNo.Yes.equals(respondingParty.getValue().getUser().getSolicitorRepresented())
-                    && respondingParty.getValue().getResponse() != null) {
-                    final boolean hasSubmitted = YesOrNo.Yes.equals(respondingParty.getValue().getResponse().getC7ResponseSubmitted());
-                    String representedRespondentName = respondingParty.getValue().getLabelForDynamicList();
-                    if (hasSubmitted) {
-                        return respondentSolicitorTaskListRenderer
-                            .render(
-                                null,
-                                null,
-                                respondent,
-                                representedRespondentName,
-                                hasSubmitted,
-                                caseData.getId()
-                            );
-                    } else {
-                        final List<RespondentTask> tasks = taskListService.getRespondentSolicitorTasks(respondingParty.getValue());
-
-                        List<RespondentEventValidationErrors> eventErrors = respondentTaskErrorService.getEventErrors();
-
-                        List<RespondentSolicitorEvents> events = taskListService.getRespondentsEvents();
-                        eventErrors.removeIf(e -> !events.contains(e.getEvent()));
-                        return respondentSolicitorTaskListRenderer
-                            .render(
-                                tasks,
-                                eventErrors,
-                                respondent,
-                                representedRespondentName,
-                                hasSubmitted,
-                                caseData.getId()
-                            );
-                    }
-                }
-            }
-        }
-        return respondentTaskList;
     }
 }
